Initial 2D implementation and bugfix
Adds an initial version of Grid Sensor for 2D and fixes a bug with Area2d nodes not getting created in editor upon the sensor being added on scene.

There is still a potential issue in case where nodes are edited manually before saving a scene. Then the grid displayed in editor will not align with the actual grid once the application is launched unless a parameter is changed in the inspector which would rebuild the grid.